prelude/pyproject.toml
[project]
name = "prelude"
version = "0.1.0"
description = "RAG + Minimal Agent experimental playground"
requires-python = ">=3.11"
dependencies = [
    "playwright",
    "beautifulsoup4",
    "markdownify",
    "tiktoken",
    "sentence-transformers",
    "chromadb",
    "anthropic",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "rag"]
